requirements.txt
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
confluent-kafka==2.0.2
psycopg2-binary==2.9.6
requests==2.28.1
SQLAlchemy==2.0.0
python-dotenv==0.21.0
plotly==5.10.0
pytz==2023.3
keyboard==0.13.5